fix: Fix for providers
