+++ languages/cpp/src/shared/src/Gateway/unidi/gateway_impl.h
@@ -60,6 +60,11 @@ namespace FireboltSDK
             return transport->Invoke(method, parameters, response);
         }
 
+        Firebolt::Error Response(unsigned id, const std::string &method, const JsonObject &response)
+        {
+            return Firebolt::Error::General;
+        }
+
         template <typename RESPONSE>
         Firebolt::Error Subscribe(const string& event, const string& parameters, RESPONSE& response)
         {
@@ -77,8 +82,8 @@ namespace FireboltSDK
             return transport->Unsubscribe(event, parameters);
         }
 
-        template <typename RESPONSE, typename CALLBACK>
-        Firebolt::Error RegisterProviderInterface(const std::string &capability, const std::string &interface, const std::string &method, const JsonObject &parameters, const CALLBACK& callback)
+        template <typename RESPONSE, typename PARAMETERS, typename CALLBACK>
+        Firebolt::Error RegisterProviderInterface(const std::string &method, const PARAMETERS &parameters, const CALLBACK& callback, void* usercb)
         {
             return Firebolt::Error::General;
         }
